Disable MOZ_BLOCK_PROFILE_DOWNGRADE.
It would be desirable to prevent downgrades from one major version of
IceCat to another.  However, as MOZ_BLOCK_PROFILE_DOWNGRADE is
currently implemented, it prevents downgrades from one build of
IceCat-68.2 to an earlier build of IceCat-68.2.  Until we can find a
satisfactory solution, we disable this functionality entirely.

* data/patches/moz-configure-changes.patch: Set
MOZ_BLOCK_PROFILE_DOWNGRADE to 'False'.