Break makeicecat down into several smaller functions
* makeicecat: Break down into a series of smaller functions, improving
the organization and manageability of the script, making it easier to
make certain steps idempotent, validatable, etc.  Use a more portable
'#!/usr/bin/env bash' shebang, useful e.g. when building on FreeBSD.
Add a DEVEL variable to do or not do a few things depending on whether
it is set, for a better/faster development experience.

Co-authored-by: Amin Bandali <bandali@gnu.org>